version-2: modify some warns: 1) Multiple versions of scala libraries detected! -- caused due to a certain compillence dependency of spark-core, details to notes in github 2) Using platform encoding (UTF-8 actually) to copy filtered resources

+++ recommender/StatisticsRecommender/src/main/scala/com/lyu/statistic/StatisticsRecommender.scala
@@ -14,7 +14,7 @@ import org.apache.spark.sql.{DataFrame, SparkSession}
  * @param score
  * @param timestamp
  */
-case class Rating( userId: Int, productId: Int, score: Double, timestamp: Int )
+case class Rating( userId: Int, productId: Int, score: Double, timestamp: BigInt )
 
 /**
  * MongoDB 连接配置 样例类
